+++ CHANGELOG.md
@@ -12,6 +12,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [3.4.0] 2020-06-10
+### Added
+- Added support for Application Notifications with object `plank.NotificationsType` 
+  - `plank.UpdateApplicationNotifications` to create, update app notifications 
+  - `plank.GetApplicationNotifications` to read app notifications
+  - `plank.FillAppNotificationFields` to fill default values for them
+  - `plank.ValidateAppNotification` to validate struct
+
+### Changed
+
+### Fixed
+
+### Removed
 ## [3.3.0] 2020-06-08
 
 ### Added
@@ -131,6 +144,7 @@ paylods from 4xx and 5xx responses in the `plank.FailedResponse` struct.
   struct makes sense for the context.
 
 [Unreleased]: https://github.com/armory/plank/compare/v1.3.0...HEAD
+[3.4.0]: https://github.com/armory/plank/compare/v3.3.0...v3.4.0
 [3.3.0]: https://github.com/armory/plank/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/armory/plank/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/armory/plank/compare/v3.0.0...v3.1.0